Adds a global flag around legacy support for IE.
The micro `clearfix` and `inline-block` mixins contain some pretty handy fixes for IE 8 and less, but not really needed in modern browsers. This pull request just adds a global flag whether to output the IE hacks or not. This can drastically cut down on the CSS output within a project. 

While it'd be great to fully remove these hacks, I believe, having them remain available is way too useful still.