Fix `font-url` for `sass-rails` > 5.x
When using newer versions of `sass-rails`, the `$asset-pipeline` flag
wasn't working properly. Although the `font-url` asset helper from
`sass-rails` was being generated into the resulting CSS, it was not
being post-processed by the asset pipeline.

The way we were previously outputting the `font-url` caused `sass-rails`
to not recognize the `font-url` helper. This change appends the helper
in a more direct way, and this resolves the issue.

Fixes #603