Add deprecate mixins
This borrows heavily from how we deprecate features in Neat using Sass @warn:
https://github.com/thoughtbot/neat/commit/5c337cc8344d9632f0d2fb41cc4c04fc9d665d9b

One key difference from Neat is that we only want to use this mixin for
deprecations, not all warnings. The reason is because with CSS, it's not always
a straightforward task to upgrade libraries or frameworks; particularly major
releases. So we offer a toggle to globally disable deprecation warnings, but for
other meaningful warnings we need to throw, we still output those.

To deprecate a feature, you include the `_bourbon-deprecate` mixin to throw a
Sass warning, providing a detailed message that gives context.

We offer a variable to allow people to disable deprecation warnings from being
output.

One problem with this system is that we cannot use these deprecation mixins to
deprecate a function. This is because Sass doesn't allow you to include a mixin
in a function. For now, since we only have a few functions up for deprecation,
we can live with this and manually write warnings for those.