Document the governance of this project
This documents how users, reviews, contributors, and maintainers are
expected to interact with each other. This is separate from a code of
conduct in that it describes how the project is intended to govern, and
not how individuals are expected to interact with each other as humans.

This initial documentation describes how it currently functions. We can
make further improvements from here.