Fix flaky tests
Different implementations and/or versions of Sass appear to handle
decimals differently, and so our tests that checked against number
values with decimals, we would see failures because the output would
vary in how many decimal places were output.

The `--precision` flag can be passed into the Sass CLI to tell Sass how
many digits of precision to use.

https://sass-lang.com/documentation/cli/ruby-sass#precision