Refactor `font-face` mixin, improve error handling
This refactors the `font-face` mixin so that we can throw a proper and
descriptive error when someone attempts to use it with a font file
format that we do not support (e.g. OTF).

To do so, we change how we handle the mapping of the font file formats
to the `src` that we output. Instead of iterating over our supported
formats map and then checking it against the list that the user passes
into the `$file-formats` argument, we reverse that and work from the
user's input and map it against our `$formats-map`. This allows us to
throw an `@error` if the user expects output from an unsupported format.

Related: https://github.com/thoughtbot/bourbon/issues/1080