Use CSS revert in hide-visually mixin
When the `unhide` value is passed into the `hide-visually` mixin, we
"undo" the CSS we initially set in the mixin by setting default values.
The `revert` value can be used more appropriately here as it will reset
these property values to the cascaded value. Since `revert` is not
fully supported, we wrap its use in an `@supports` query.

https://developer.mozilla.org/en-US/docs/Web/CSS/revert