Upgrade to Ruby 2.7 (#1113)
* Update to Ruby latest 2.6 version

As part of 2.7 upgrade, we want to update to the latest 2.6 Ruby version
in order to see the latest deprecation warnings.

This commit does this upgrade.

* Upgrade Ruby to 2.7.6

As we didn't have any deprecation warning in the previous commit - which
upgraded Ruby to the latest 2.6 version, it should be safe to upgrade it
to 2.7.

Future work needs to be done to address any new deprecation warning that
arises with this upgrade.