Upgrade gems to better work w/ ruby 3.1
As we are upgrading our ruby version, there are dependencies
that need to be updated as well to ensure that their are
behaving as expected.

The following gems were updated:
- `bundler`: v2.2.33 doesn't work well with ruby v3.1 and
shows the following warning:
  - `Calling
  'DidYouMean::SPELL_CHECKERS.merge!(error_name => spell_checker)'
  has been deprecated. Please call
  'DidYouMean.correct_error(error_name, spell_checker)' instead.
- `contracts`: according to the warning during bundling,
v0.16 should be used only with ruby v2.x. For ruby v3.x +, we
should use v0.17
- `cucumber`: v2.0 has some warnings about redefining methods
that already exists in the gem.